Ask Ubu Upgrade
I just rewrote the 'Ask Ubu' code to wake up our master when you click
his green button, and try to keep his loud antics down to a short
amount of time.

I would like to spin his spiral gidouille independently from the head
shaking, so the wheel can spin longer, like for 30 seconds, without
winding up and down every few seconds, which could drive some folks on
edge. I doubt that I wrote the right code for that, so I may consult
with you to see if there’s a simple way to do that.

I borrowed a lot of that code from the Bali Cuckoo Clock, so the Ubu is
now going through 7 different stages on his path to redemption, to
separate out some of these functions so they don’t all happen at once,
as loosely tracked by the ubuState variable.

Eventually, we may also have him request a sound from the server, but
for now I commented out all this code, because it’s complicated enough
as it is. Plus, the Adafruit MotorShield also uses i2C to communicate
to the motors, so it seems best to defer that complexity for later.