Updated sound software
Updated sound software for most of the wonderboxes that are ready for
it, both on Github and on their Arduinos or Diavolinos. This includes
new code for Eros, Other Dimensions, Pay Attention and Ubu Roi Stage,
which has been installed on these boxes.  The latest status of each box
is indicated in columns D and E on this Wonderbox spreadsheet:
https://docs.google.com/spreadsheets/d/13zPk_wpKfXi6CO7xE0nm0pAtctWqHBnh
E1YoH1_kNxw/edit#gid=0