Updated sound playback code
Updated sound client code for the X-Ray Skull wonderbox: the delay is
now equal to 200 milliseconds multiplied by the box number, so the
boxes don’t all call the sound server at the same time. Also added some
code to detach the servo when it is not in use, to avoid straining it.